+++ tests/api/test_account.py
@@ -9,7 +9,6 @@ from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import ValidationError
 from django.core.files import File
 from django.core.validators import URLValidator
-from django.test import override_settings
 from freezegun import freeze_time
 from prices import Money
 
@@ -305,6 +304,18 @@ def test_user_query_permission_manage_users_get_customer(
     assert customer_user.email == data["email"]
 
 
+def test_user_query_as_service_account(
+    service_account_api_client, customer_user, permission_manage_users, service_account
+):
+    service_account.permissions.add(permission_manage_users)
+    customer_id = graphene.Node.to_global_id("User", customer_user.pk)
+    variables = {"id": customer_id}
+    response = service_account_api_client.post_graphql(USER_QUERY, variables)
+    content = get_graphql_content(response)
+    data = content["data"]["user"]
+    assert customer_user.email == data["email"]
+
+
 def test_user_query_permission_manage_users_get_staff(
     staff_api_client, staff_user, permission_manage_users
 ):
@@ -579,9 +590,7 @@ ACCOUNT_REGISTER_MUTATION = """
 """
 
 
-@override_settings(ENABLE_ACCOUNT_CONFIRMATION_BY_EMAIL=True)
-@patch("saleor.account.emails._send_account_confirmation_email")
-def test_customer_register(send_account_confirmation_email_mock, user_api_client):
+def test_customer_register(user_api_client):
     email = "customer@example.com"
     variables = {"email": email, "password": "Password"}
     query = ACCOUNT_REGISTER_MUTATION
@@ -590,7 +599,6 @@ def test_customer_register(send_account_confirmation_email_mock, user_api_client
     content = get_graphql_content(response)
     data = content["data"][mutation_name]
     assert not data["errors"]
-    assert send_account_confirmation_email_mock.delay.call_count == 1
     new_user = User.objects.get(email=email)
 
     response = user_api_client.post_graphql(query, variables)
@@ -603,16 +611,6 @@ def test_customer_register(send_account_confirmation_email_mock, user_api_client
     customer_creation_event = account_events.CustomerEvent.objects.get()
     assert customer_creation_event.type == account_events.CustomerEvents.ACCOUNT_CREATED
     assert customer_creation_event.user == new_user
-
-
-@override_settings(ENABLE_ACCOUNT_CONFIRMATION_BY_EMAIL=False)
-@patch("saleor.account.emails._send_account_confirmation_email")
-def test_customer_register_disabled_email_confirmation(
-    send_account_confirmation_email_mock, user_api_client
-):
-    variables = {"email": "customer@example.com", "password": "Password"}
-    user_api_client.post_graphql(ACCOUNT_REGISTER_MUTATION, variables)
-    assert send_account_confirmation_email_mock.delay.call_count == 0
 
 
 CUSTOMER_CREATE_MUTATION = """
@@ -2111,18 +2109,6 @@ REQUEST_PASSWORD_RESET_MUTATION = """
 """
 
 
-CONFIRM_ACCOUNT_MUTATION = """
-    mutation ConfirmAccount($email: String!, $token: String!) {
-        confirmAccount(email: $email, token: $token) {
-            errors {
-                field
-                message
-            }
-        }
-    }
-"""
-
-
 @patch("saleor.account.emails._send_password_reset_email")
 def test_account_reset_password(
     send_password_reset_email_mock, user_api_client, customer_user
@@ -2139,41 +2125,6 @@ def test_account_reset_password(
     url = send_password_reset_email_mock.mock_calls[0][1][1]
     url_validator = URLValidator()
     url_validator(url)
-
-
-def test_account_confirmation(user_api_client, customer_user):
-    customer_user.is_active = False
-    customer_user.save()
-
-    variables = {
-        "email": customer_user.email,
-        "token": default_token_generator.make_token(customer_user),
-    }
-    user_api_client.post_graphql(CONFIRM_ACCOUNT_MUTATION, variables)
-
-    customer_user.refresh_from_db()
-    assert customer_user.is_active is True
-
-
-def test_account_confirmation_invalid_user(user_api_client, customer_user):
-    variables = {
-        "email": "non-existing@nope.com",
-        "token": default_token_generator.make_token(customer_user),
-    }
-    response = user_api_client.post_graphql(CONFIRM_ACCOUNT_MUTATION, variables)
-    content = get_graphql_content(response)
-    assert content["data"]["confirmAccount"]["errors"] == [
-        {"field": "email", "message": "User with this email doesn't exist"}
-    ]
-
-
-def test_account_confirmation_invalid_token(user_api_client, customer_user):
-    variables = {"email": customer_user.email, "token": "invalid_token"}
-    response = user_api_client.post_graphql(CONFIRM_ACCOUNT_MUTATION, variables)
-    content = get_graphql_content(response)
-    assert content["data"]["confirmAccount"]["errors"] == [
-        {"field": "token", "message": "Invalid or expired token."}
-    ]
 
 
 @patch("saleor.account.emails._send_password_reset_email")
@@ -3099,27 +3050,14 @@ def test_address_query_as_not_owner(
     assert not data
 
 
-def test_address_query_as_service_account_with_permission(
-    service_account_api_client,
-    service_account,
-    address_other_country,
-    permission_manage_users,
+def test_address_query_as_service_account(
+    service_account_api_client, service_account, address_other_country
 ):
-    service_account.permissions.add(permission_manage_users)
     variables = {"id": graphene.Node.to_global_id("Address", address_other_country.pk)}
     response = service_account_api_client.post_graphql(ADDRESS_QUERY, variables)
     content = get_graphql_content(response)
     data = content["data"]["address"]
     assert data["country"]["code"] == address_other_country.country.code
-
-
-def test_address_query_as_service_account_without_permission(
-    service_account_api_client, service_account, address_other_country
-):
-
-    variables = {"id": graphene.Node.to_global_id("Address", address_other_country.pk)}
-    response = service_account_api_client.post_graphql(ADDRESS_QUERY, variables)
-    assert_no_permission(response)
 
 
 def test_address_query_as_anonymous_user(api_client, address_other_country):